% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.alpha.r
\name{add.alpha}
\alias{add.alpha}
\title{Add transparency to colors}
\usage{
add.alpha(col = "black", alpha = 1)
}
\arguments{
\item{col}{Base color(s)}

\item{alpha}{Transparency of the colors (between 0 and 1 - 1 is opaque)}
}
\description{
Add transparency to colors
}

