% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.process.r
\name{bernoulli.process}
\alias{bernoulli.process}
\title{Interarrival times and arrival times of a poisson process}
\usage{
bernoulli.process(probability = 0.5, events = NULL, time.running = NULL)
}
\arguments{
\item{probability}{The probability of success in the poisson process}

\item{events}{The number of events in the process (or \code{NULL} if defining
the cut off point by the time)}

\item{time.running}{The amount of time for the process to run (or \code{NULL}
if defining the cut off by the number of events)}
}
\value{
An numeric matrix with 2 columns. The first column contains the interarrival
  times and the second contains the arrival times.
}
\description{
Interarrival times and arrival times of a poisson process
}

