% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.urn.r
\name{draw.urn}
\alias{draw.urn}
\title{Results from drawing different color balls from an urn}
\usage{
draw.urn(n.iter = 1e+05, n.colors = 2, n.balls = c(10, 10), n.drawn = 5,
  replacement = FALSE, collapse = FALSE, parallel = TRUE)
}
\arguments{
\item{n.iter}{The number of iterations in the monte carlo simulation (integer > 0)}

\item{n.colors}{The number of different colored balls (integer > 0)}

\item{n.balls}{The count of each color balls (vector of integers with same length as \code{n.colors})}

\item{n.drawn}{The number of balls drawn from the urn in each iteration}

\item{replacement}{Are the balls placed back in the urn after they're drawn? (defaults to \code{FALSE})}

\item{collapse}{Should the return be the draws or the count of each color? (defaults to \code{FALSE})}

\item{parallel}{Should the simulation take advantage of multiple cores (defaults to \code{TRUE} -
highly recommended if \code{n.iter > 1e5})}
}
\value{
If \code{collapse == TRUE} the return is a \code{n.iter x n.colors} integer matrix
  where each row is the count of the colors of balls drawn. Otherwise, the return is a
  \code{n.iter x n.drawn} integer matrix where each row is the order that colors were drawn in.
}
\description{
Results from drawing different color balls from an urn
}

