% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.matrix.r
\name{latex.matrix}
\alias{latex.matrix}
\title{Used to quickly output LaTeX code for a matrix in R.}
\usage{
latex.matrix(mat, in_mm = TRUE)
}
\arguments{
\item{mat}{$A$, a matrix in R}

\item{in_mm}{Boolean for whether or not include an entering and exiting of
math mode}
}
\value{
Latex code for the matrix $A$
}
\description{
Used to quickly output LaTeX code for a matrix in R.
}

