% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.cor.r
\name{panel.cor}
\alias{panel.cor}
\title{Used to replace duplicate scatterplots in a matrix}
\usage{
panel.cor(x, y, digits = 2, cex.cor, include.p = FALSE, ...)
}
\description{
Used to replace duplicate scatterplots in a matrix
}

