% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.dice.r
\name{roll.dice}
\alias{roll.dice}
\title{Results of rolling dice}
\usage{
roll.dice(n.iter = 1e+05, n.dice = 1, sides = 6, sum.dice = FALSE,
  parallel = TRUE)
}
\arguments{
\item{n.iter}{The number of iterations in the monte carlo simulation (integer > 0)}

\item{n.dice}{The number of dice that should be rolled on each trial (integer > 0)}

\item{sides}{The number of sides on each die (integer > 0)}

\item{sum.dice}{Should the result be the sum of the dice on each iteration (defaults to \code{FALSE})}

\item{parallel}{Should the simulation take advantage of multiple cores (defaults to \code{TRUE} -
highly recommended if \code{n.iter > 1e5})}
}
\value{
If \code{sum.dice == TRUE} then the return is a vector of \code{n.iter} integer entries
  else the return is a \code{n.iter x n.dice} integer matrix
}
\description{
Results of rolling dice
}

